moved a function to tensorrt_llm_run file
